+++ previous-versions/sync-for-expenses-version-1/src/codatsyncexpenses/sdkconfiguration.py
@@ -20,9 +20,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = 'prealpha'
-    sdk_version: str = '0.3.0'
-    gen_version: str = '2.169.0'
-    user_agent: str = 'speakeasy-sdk/python 0.3.0 2.169.0 prealpha codat-sync-for-expenses-version-1'
+    sdk_version: str = '0.3.1'
+    gen_version: str = '2.173.0'
+    user_agent: str = 'speakeasy-sdk/python 0.3.1 2.173.0 prealpha codat-sync-for-expenses-version-1'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> Tuple[str, Dict[str, str]]: